Fix component(s) insertion regression when it exists
A component(s) which already exist on an entity is
not inserted.This was because the entity is moved
from the table into the same table, this logic is
unsound as it is moved to a new row invalidting
the previous row thus the entity should not be moved.
Instead,just update the entity at its row.